Stay out of utils, add comments about bandit error
- Even if we try to determine a bucket region in two passes (with and
  without a client), the clientless piece is not generic enough to
  justify moving it to utils.

- The bandit error about auditing urlopen makes sense generally,
  but it looks like we've got it covered in this case. Add an
  exclusion comment (for bandit) and explanatory note (for future
  maintainers).
